requirements.txt
aiogram==2.16
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
filelock==3.4.0
frozenlist==1.2.0
huggingface-hub==0.1.2
idna==3.3
joblib==1.1.0
lyricsgenius==3.0.1
multidict==5.2.0
numpy==1.21.4
packaging==21.2
pandas==1.3.4
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
sacremoses==0.0.46
six==1.16.0
soupsieve==2.3.1
textwrap3==0.9.2
tokenizers==0.10.3
tqdm==4.62.3
transformers==4.12.4
typing-extensions==4.0.0
urllib3==1.26.7
yarl==1.7.2